@@ -96,7 +96,7 @@ Bob calls `func`. As a result, `owner` is overridden to `0`.
 
         for contract in self.compilation_unit.contracts:
             for function in contract.functions:
-                if function.is_implemented:
+                if function.is_implemented and function.entry_point:
                     uninitialized_storage_variables = [
                         v for v in function.local_variables if v.is_storage and v.uninitialized
                     ]
