@@ -6,6 +6,7 @@ from slither.core.declarations.custom_error_contract import CustomErrorContract
 from slither.core.declarations.custom_error_top_level import CustomErrorTopLevel
 from slither.core.declarations.function_contract import FunctionContract
 from slither.core.expressions.literal import Literal
+from slither.core.solidity_types import TypeAlias
 from slither.core.solidity_types.array_type import ArrayType
 from slither.core.solidity_types.elementary_type import (
     ElementaryType,
@@ -111,7 +112,7 @@ def _find_from_type_name(  # pylint: disable=too-many-locals,too-many-branches,t
     if not var_type:
         if name.startswith("function "):
             found = re.findall(
-                "function \(([ ()\[\]a-zA-Z0-9\.,]*?)\)(?: payable)?(?: (?:external|internal|pure|view))?(?: returns \(([a-zA-Z0-9() \.,]*)\))?",
+                r"function \(([ ()\[\]a-zA-Z0-9\.,]*?)\)(?: payable)?(?: (?:external|internal|pure|view))?(?: returns \(([a-zA-Z0-9() \.,]*)\))?",
                 name,
             )
             assert len(found) == 1
@@ -158,10 +159,10 @@ def _find_from_type_name(  # pylint: disable=too-many-locals,too-many-branches,t
         if name.startswith("mapping("):
             # nested mapping declared with var
             if name.count("mapping(") == 1:
-                found = re.findall("mapping\(([a-zA-Z0-9\.]*) => ([ a-zA-Z0-9\.\[\]]*)\)", name)
+                found = re.findall(r"mapping\(([a-zA-Z0-9\.]*) => ([ a-zA-Z0-9\.\[\]]*)\)", name)
             else:
                 found = re.findall(
-                    "mapping\(([a-zA-Z0-9\.]*) => (mapping\([=> a-zA-Z0-9\.\[\]]*\))\)",
+                    r"mapping\(([a-zA-Z0-9\.]*) => (mapping\([=> a-zA-Z0-9\.\[\]]*\))\)",
                     name,
                 )
             assert len(found) == 1
@@ -204,7 +205,7 @@ def _add_type_references(type_found: Type, src: str, sl: "SlitherCompilationUnit
 def parse_type(
     t: Union[Dict, UnknownType],
     caller_context: Union[CallerContextExpression, "SlitherCompilationUnitSolc"],
-):
+) -> Type:
     """
     caller_context can be a SlitherCompilationUnitSolc because we recursively call the function
     and go up in the context's scope. If we are really lost we just go over the SlitherCompilationUnitSolc
@@ -229,6 +230,8 @@ def parse_type(
     from slither.solc_parsing.variables.top_level_variable import TopLevelVariableSolc
 
     sl: "SlitherCompilationUnit"
+    renaming: Dict[str, str]
+    user_defined_types: Dict[str, TypeAlias]
     # Note: for convenicence top level functions use the same parser than function in contract
     # but contract_parser is set to None
     if isinstance(caller_context, SlitherCompilationUnitSolc) or (
@@ -238,10 +241,14 @@ def parse_type(
         if isinstance(caller_context, SlitherCompilationUnitSolc):
             sl = caller_context.compilation_unit
             next_context = caller_context
+            renaming = {}
+            user_defined_types = {}
         else:
             assert isinstance(caller_context, FunctionSolc)
             sl = caller_context.underlying_function.compilation_unit
             next_context = caller_context.slither_parser
+            renaming = caller_context.underlying_function.file_scope.renaming
+            user_defined_types = caller_context.underlying_function.file_scope.user_defined_types
         structures_direct_access = sl.structures_top_level
         all_structuress = [c.structures for c in sl.contracts]
         all_structures = [item for sublist in all_structuress for item in sublist]
@@ -272,10 +279,17 @@ def parse_type(
         all_structuress = [c.structures for c in scope.contracts.values()]
         all_structures = [item for sublist in all_structuress for item in sublist]
         all_structures += structures_direct_access
+
         enums_direct_access = []
-        all_enums = scope.enums.values()
+        all_enumss = [c.enums for c in scope.contracts.values()]
+        all_enums = [item for sublist in all_enumss for item in sublist]
+        all_enums += scope.enums.values()
+
         contracts = scope.contracts.values()
         functions = list(scope.functions)
+
+        renaming = scope.renaming
+        user_defined_types = scope.user_defined_types
     elif isinstance(caller_context, (ContractSolc, FunctionSolc)):
         sl = caller_context.compilation_unit
         if isinstance(caller_context, FunctionSolc):
@@ -285,9 +299,11 @@ def parse_type(
             assert isinstance(underlying_func, FunctionContract)
             contract = underlying_func.contract
             next_context = caller_context.contract_parser
+            scope = caller_context.underlying_function.file_scope
         else:
             contract = caller_context.underlying_contract
             next_context = caller_context
+            scope = caller_context.underlying_contract.file_scope
 
         structures_direct_access = contract.structures
         structures_direct_access += contract.file_scope.structures.values()
@@ -301,6 +317,9 @@ def parse_type(
         all_enums += contract.file_scope.enums.values()
         contracts = contract.file_scope.contracts.values()
         functions = contract.functions + contract.modifiers
+
+        renaming = scope.renaming
+        user_defined_types = scope.user_defined_types
     else:
         raise ParsingError(f"Incorrect caller context: {type(caller_context)}")
 
@@ -311,8 +330,13 @@ def parse_type(
         key = "name"
 
     if isinstance(t, UnknownType):
+        name = t.name
+        if name in renaming:
+            name = renaming[name]
+        if name in user_defined_types:
+            return user_defined_types[name]
         return _find_from_type_name(
-            t.name,
+            name,
             functions,
             contracts,
             structures_direct_access,
@@ -328,8 +352,13 @@ def parse_type(
 
     if t[key] == "UserDefinedTypeName":
         if is_compact_ast:
+            name = t["typeDescriptions"]["typeString"]
+            if name in renaming:
+                name = renaming[name]
+            if name in user_defined_types:
+                return user_defined_types[name]
             type_found = _find_from_type_name(
-                t["typeDescriptions"]["typeString"],
+                name,
                 functions,
                 contracts,
                 structures_direct_access,
@@ -342,8 +371,14 @@ def parse_type(
 
         # Determine if we have a type node (otherwise we use the name node, as some older solc did not have 'type').
         type_name_key = "type" if "type" in t["attributes"] else key
+
+        name = t["attributes"][type_name_key]
+        if name in renaming:
+            name = renaming[name]
+        if name in user_defined_types:
+            return user_defined_types[name]
         type_found = _find_from_type_name(
-            t["attributes"][type_name_key],
+            name,
             functions,
             contracts,
             structures_direct_access,
@@ -357,8 +392,13 @@ def parse_type(
     # Introduced with Solidity 0.8
     if t[key] == "IdentifierPath":
         if is_compact_ast:
-            type_found = _find_from_type_name(
-                t["name"],
+            name = t["name"]
+            if name in renaming:
+                name = renaming[name]
+            if name in user_defined_types:
+                return user_defined_types[name]
+            type_found =  _find_from_type_name(
+                name,
                 functions,
                 contracts,
                 structures_direct_access,
