@@ -2,6 +2,7 @@ import json
 import os
 import re
 import sys
+from pathlib import Path
 from typing import List, Dict, Tuple
 
 import pytest
@@ -59,7 +60,7 @@ VERSIONS_04 = make_version(4, 0, 26)
 VERSIONS_05 = make_version(5, 0, 17)
 VERSIONS_06 = make_version(6, 0, 12)
 VERSIONS_07 = make_version(7, 0, 6)
-VERSIONS_08 = make_version(8, 0, 12)
+VERSIONS_08 = make_version(8, 0, 15)
 
 ALL_VERSIONS = VERSIONS_04 + VERSIONS_05 + VERSIONS_06 + VERSIONS_07 + VERSIONS_08
 
@@ -309,6 +310,10 @@ ALL_TESTS = [
         "minmax-0.6.8.sol",
         make_version(6, 8, 9) + VERSIONS_07 + VERSIONS_08,
     ),
+    Test(
+        "minmax-0.8.8.sol",
+        make_version(8, 8, 15),
+    ),
     Test("dowhile-0.4.0.sol", VERSIONS_04),
     Test(
         "dowhile-0.4.5.sol",
@@ -394,6 +399,26 @@ ALL_TESTS = [
         "while-all.sol",
         ALL_VERSIONS,
     ),
+    Test(
+        "complex_imports/import_free/Caller.sol",
+        ["0.8.2"],
+    ),
+    Test("custom_error_with_state_variable.sol", make_version(8, 4, 12)),
+    Test("complex_imports/import_aliases/test.sol", VERSIONS_08),
+    # 0.8.9 crashes on our testcase
+    Test(
+        "user_defined_value_type/user_defined_types-0.8.8.sol", ["0.8.8"] + make_version(8, 10, 15)
+    ),
+    Test("user_defined_value_type/argument-0.8.8.sol", ["0.8.8"] + make_version(8, 10, 15)),
+    Test("user_defined_value_type/calldata-0.8.8.sol", ["0.8.8"] + make_version(8, 10, 15)),
+    Test("user_defined_value_type/constant-0.8.8.sol", ["0.8.8"] + make_version(8, 10, 15)),
+    Test("user_defined_value_type/erc20-0.8.8.sol", ["0.8.8"] + make_version(8, 10, 15)),
+    Test("user_defined_value_type/in_parenthesis-0.8.8.sol", ["0.8.8"] + make_version(8, 10, 15)),
+    Test("bytes_call.sol", ["0.8.12"]),
+    Test("free_functions/libraries_from_free.sol", ["0.8.12"]),
+    Test("free_functions/new_operator.sol", ["0.8.12"]),
+    Test("free_functions/library_constant_function_collision.sol", ["0.8.12"]),
+    Test("ternary-with-max.sol", ["0.8.15"]),
 ]
 # create the output folder if needed
 try:
@@ -492,6 +517,10 @@ def _generate_test(test_item: Test, skip_existing=False):
 
         actual = generate_output(sl)
         print(f"Generate {expected_file}")
+
+        # pylint: disable=no-member
+        Path(expected_file).parents[0].mkdir(parents=True, exist_ok=True)
+
         with open(expected_file, "w", encoding="utf8") as f:
             json.dump(actual, f, indent="  ")
 
@@ -517,6 +546,9 @@ def _generate_compile(test_item: Test, skip_existing=False):
         set_solc(version)
         print(f"Compiled to {expected_file}")
         cc = CryticCompile(test_file, solc_force_legacy_json=flavor == "legacy")
+
+        # pylint: disable=no-member
+        Path(expected_file).parents[0].mkdir(parents=True, exist_ok=True)
 
         save_to_zip([cc], expected_file)
 
