@@ -117,6 +117,26 @@ ALL_TEST_OBJECTS = [
         "0.4.25",
     ),
     Test(
+        all_detectors.ReentrancyReadBeforeWritten,
+        "comment.sol",
+        "0.8.2",
+    ),
+    Test(
+        all_detectors.ReentrancyReadBeforeWritten,
+        "no-reentrancy-staticcall.sol",
+        "0.5.16",
+    ),
+    Test(
+        all_detectors.ReentrancyReadBeforeWritten,
+        "no-reentrancy-staticcall.sol",
+        "0.6.11",
+    ),
+    Test(
+        all_detectors.ReentrancyReadBeforeWritten,
+        "no-reentrancy-staticcall.sol",
+        "0.7.6",
+    ),
+    Test(
         all_detectors.BooleanEquality,
         "boolean-constant-equality.sol",
         "0.4.25",
@@ -376,23 +396,23 @@ ALL_TEST_OBJECTS = [
     Test(all_detectors.LockedEther, "locked_ether.sol", "0.6.11"),
     Test(all_detectors.LockedEther, "locked_ether.sol", "0.7.6"),
     Test(
-        all_detectors.ArbitrarySend,
-        "arbitrary_send.sol",
+        all_detectors.ArbitrarySendEth,
+        "arbitrary_send_eth.sol",
         "0.4.25",
     ),
     Test(
-        all_detectors.ArbitrarySend,
-        "arbitrary_send.sol",
+        all_detectors.ArbitrarySendEth,
+        "arbitrary_send_eth.sol",
         "0.5.16",
     ),
     Test(
-        all_detectors.ArbitrarySend,
-        "arbitrary_send.sol",
+        all_detectors.ArbitrarySendEth,
+        "arbitrary_send_eth.sol",
         "0.6.11",
     ),
     Test(
-        all_detectors.ArbitrarySend,
-        "arbitrary_send.sol",
+        all_detectors.ArbitrarySendEth,
+        "arbitrary_send_eth.sol",
         "0.7.6",
     ),
     Test(
@@ -636,6 +656,16 @@ ALL_TEST_OBJECTS = [
         "0.5.16",
     ),
     Test(
+        all_detectors.ShadowingAbstractDetection,
+        "shadowing_state_variable.sol",
+        "0.7.5",
+    ),
+    Test(
+        all_detectors.ShadowingAbstractDetection,
+        "public_gap_variable.sol",
+        "0.7.5",
+    ),
+    Test(
         all_detectors.StateShadowing,
         "shadowing_state_variable.sol",
         "0.4.25",
@@ -649,6 +679,16 @@ ALL_TEST_OBJECTS = [
         all_detectors.StateShadowing,
         "shadowing_state_variable.sol",
         "0.6.11",
+    ),
+    Test(
+        all_detectors.StateShadowing,
+        "shadowing_state_variable.sol",
+        "0.7.5",
+    ),
+    Test(
+        all_detectors.StateShadowing,
+        "public_gap_variable.sol",
+        "0.7.5",
     ),
     Test(
         all_detectors.StateShadowing,
@@ -820,12 +860,22 @@ ALL_TEST_OBJECTS = [
     ),
     Test(
         all_detectors.UnprotectedUpgradeable,
+        "whitelisted.sol",
+        "0.4.25",
+    ),
+    Test(
+        all_detectors.UnprotectedUpgradeable,
         "Buggy.sol",
         "0.5.16",
     ),
     Test(
         all_detectors.UnprotectedUpgradeable,
         "Fixed.sol",
+        "0.5.16",
+    ),
+    Test(
+        all_detectors.UnprotectedUpgradeable,
+        "whitelisted.sol",
         "0.5.16",
     ),
     Test(
@@ -836,6 +886,11 @@ ALL_TEST_OBJECTS = [
     Test(
         all_detectors.UnprotectedUpgradeable,
         "Fixed.sol",
+        "0.6.11",
+    ),
+    Test(
+        all_detectors.UnprotectedUpgradeable,
+        "whitelisted.sol",
         "0.6.11",
     ),
     Test(
@@ -846,6 +901,11 @@ ALL_TEST_OBJECTS = [
     Test(
         all_detectors.UnprotectedUpgradeable,
         "Fixed.sol",
+        "0.7.6",
+    ),
+    Test(
+        all_detectors.UnprotectedUpgradeable,
+        "whitelisted.sol",
         "0.7.6",
     ),
     Test(
@@ -1158,6 +1218,122 @@ ALL_TEST_OBJECTS = [
         "0.8.0",
     ),
     Test(
+        all_detectors.ShiftParameterMixup,
+        "shift_parameter_mixup.sol",
+        "0.4.25",
+    ),
+    Test(
+        all_detectors.ShiftParameterMixup,
+        "shift_parameter_mixup.sol",
+        "0.5.16",
+    ),
+    Test(
+        all_detectors.ShiftParameterMixup,
+        "shift_parameter_mixup.sol",
+        "0.6.11",
+    ),
+    Test(
+        all_detectors.ShiftParameterMixup,
+        "shift_parameter_mixup.sol",
+        "0.7.6",
+    ),
+    Test(
+        all_detectors.MissingInheritance,
+        "unimplemented_interface.sol",
+        "0.4.25",
+    ),
+    Test(
+        all_detectors.MissingInheritance,
+        "unimplemented_interface.sol",
+        "0.5.16",
+    ),
+    Test(
+        all_detectors.MissingInheritance,
+        "unimplemented_interface.sol",
+        "0.6.11",
+    ),
+    Test(
+        all_detectors.MissingInheritance,
+        "unimplemented_interface.sol",
+        "0.7.6",
+    ),
+    # Does not work on the CI. Most likely because of solc 0.4.2?
+    # Test(
+    #     all_detectors.EnumConversion,
+    #     "enum_conversion.sol",
+    #     "0.4.2",
+    # ),
+    Test(
+        all_detectors.MultipleConstructorSchemes,
+        "multiple_constructor_schemes.sol",
+        "0.4.22",
+    ),
+    Test(
+        all_detectors.DeprecatedStandards,
+        "deprecated_calls.sol",
+        "0.4.25",
+    ),
+    Test(
+        all_detectors.DivideBeforeMultiply,
+        "divide_before_multiply.sol",
+        "0.4.25",
+    ),
+    Test(
+        all_detectors.DivideBeforeMultiply,
+        "divide_before_multiply.sol",
+        "0.5.16",
+    ),
+    Test(
+        all_detectors.DivideBeforeMultiply,
+        "divide_before_multiply.sol",
+        "0.6.11",
+    ),
+    Test(
+        all_detectors.DivideBeforeMultiply,
+        "divide_before_multiply.sol",
+        "0.7.6",
+    ),
+    Test(
+        all_detectors.TypeBasedTautology,
+        "type_based_tautology.sol",
+        "0.4.25",
+    ),
+    Test(
+        all_detectors.TypeBasedTautology,
+        "type_based_tautology.sol",
+        "0.5.16",
+    ),
+    Test(
+        all_detectors.TypeBasedTautology,
+        "type_based_tautology.sol",
+        "0.6.11",
+    ),
+    Test(
+        all_detectors.TypeBasedTautology,
+        "type_based_tautology.sol",
+        "0.7.6",
+    ),
+    Test(
+        all_detectors.SimilarVarsDetection,
+        "similar_variables.sol",
+        "0.4.25",
+    ),
+    Test(
+        all_detectors.SimilarVarsDetection,
+        "similar_variables.sol",
+        "0.5.16",
+    ),
+    Test(
+        all_detectors.SimilarVarsDetection,
+        "similar_variables.sol",
+        "0.6.11",
+    ),
+    Test(
+        all_detectors.SimilarVarsDetection,
+        "similar_variables.sol",
+        "0.7.6",
+    ),
+    Test(
         all_detectors.MsgValueInLoop,
         "msg_value_loop.sol",
         "0.4.25",
@@ -1205,6 +1381,66 @@ ALL_TEST_OBJECTS = [
     Test(
         all_detectors.DelegatecallInLoop,
         "delegatecall_loop.sol",
+        "0.8.0",
+    ),
+    Test(
+        all_detectors.ProtectedVariables,
+        "comment.sol",
+        "0.8.2",
+    ),
+    Test(
+        all_detectors.ArbitrarySendErc20NoPermit,
+        "arbitrary_send_erc20.sol",
+        "0.4.25",
+    ),
+    Test(
+        all_detectors.ArbitrarySendErc20NoPermit,
+        "arbitrary_send_erc20.sol",
+        "0.5.16",
+    ),
+    Test(
+        all_detectors.ArbitrarySendErc20NoPermit,
+        "arbitrary_send_erc20.sol",
+        "0.6.11",
+    ),
+    Test(
+        all_detectors.ArbitrarySendErc20NoPermit,
+        "arbitrary_send_erc20.sol",
+        "0.7.6",
+    ),
+    Test(
+        all_detectors.ArbitrarySendErc20NoPermit,
+        "arbitrary_send_erc20.sol",
+        "0.8.0",
+    ),
+    Test(
+        all_detectors.ArbitrarySendErc20NoPermit,
+        "arbitrary_send_erc20_inheritance.sol",
+        "0.8.0",
+    ),
+    Test(
+        all_detectors.ArbitrarySendErc20Permit,
+        "arbitrary_send_erc20_permit.sol",
+        "0.4.25",
+    ),
+    Test(
+        all_detectors.ArbitrarySendErc20Permit,
+        "arbitrary_send_erc20_permit.sol",
+        "0.5.16",
+    ),
+    Test(
+        all_detectors.ArbitrarySendErc20Permit,
+        "arbitrary_send_erc20_permit.sol",
+        "0.6.11",
+    ),
+    Test(
+        all_detectors.ArbitrarySendErc20Permit,
+        "arbitrary_send_erc20_permit.sol",
+        "0.7.6",
+    ),
+    Test(
+        all_detectors.ArbitrarySendErc20Permit,
+        "arbitrary_send_erc20_permit.sol",
         "0.8.0",
     ),
 ]
@@ -1249,11 +1485,10 @@ def test_detector(test_item: Test):
 
     for additional_file in test_item.additional_files:
         additional_path = str(pathlib.Path(test_dir_path, additional_file).absolute())
-        results_as_string = results_as_string.replace(
-            additional_path, str(pathlib.Path(GENERIC_PATH))
-        )
-    results_as_string = results_as_string.replace(test_file_path, str(pathlib.Path(GENERIC_PATH)))
-
+        additional_path = additional_path.replace("\\", "\\\\")
+        results_as_string = results_as_string.replace(additional_path, GENERIC_PATH)
+    test_file_path = test_file_path.replace("\\", "\\\\")
+    results_as_string = results_as_string.replace(test_file_path, GENERIC_PATH)
     results = json.loads(results_as_string)
 
     diff = DeepDiff(results, expected_result, ignore_order=True, verbose_level=2)
@@ -1293,13 +1528,13 @@ def _generate_test(test_item: Test, skip_existing=False):
     results = sl.run_detectors()
 
     results_as_string = json.dumps(results)
-    results_as_string = results_as_string.replace(test_file_path, str(pathlib.Path(GENERIC_PATH)))
+    test_file_path = test_file_path.replace("\\", "\\\\")
+    results_as_string = results_as_string.replace(test_file_path, GENERIC_PATH)
 
     for additional_file in test_item.additional_files:
         additional_path = str(pathlib.Path(test_dir_path, additional_file).absolute())
-        results_as_string = results_as_string.replace(
-            additional_path, str(pathlib.Path(GENERIC_PATH))
-        )
+        additional_path = additional_path.replace("\\", "\\\\")
+        results_as_string = results_as_string.replace(additional_path, GENERIC_PATH)
 
     results = json.loads(results_as_string)
     with open(expected_result_path, "w", encoding="utf8") as f:
